requirements.txt
aiogram
python-dotenv
netmiko
transliterate
pythonping
requests
pyodbc
docker